moved browse folder functionality to a new file
